+++ lib/routes.dart
@@ -19,9 +19,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:picos/screens/add_family_member_screen.dart';
 import 'package:picos/screens/add_medication_screen/add_medication_screen.dart';
 import 'package:picos/screens/add_physician_screen.dart';
-import 'package:picos/screens/configuration_screen/configuration_finished_screen.dart';
-import 'package:picos/screens/configuration_screen/configuration_pages.dart';
-import 'package:picos/screens/configuration_screen/configuration_screen.dart';
+import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_finished_screen.dart';
+import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_pages.dart';
+import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_screen.dart';
 import 'package:picos/screens/family_members_screen.dart';
 import 'package:picos/screens/main_screen/bottom_bar.dart';
 import 'package:picos/screens/my_medications_screen/my_medications_screen.dart';
