+++ lib/screens/main_screen/main_screen.dart
@@ -21,7 +21,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:picos/repository/medications_repository.dart';
 import 'package:picos/screens/login_screen.dart';
-import 'package:picos/screens/configuration_screen/configuration_screen.dart';
+import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_screen.dart';
 import 'package:picos/state/medications_list_bloc.dart';
 import 'package:picos/themes/global_theme.dart';
