+++ lib/screens/study_nurse_screen/configuration_screen/pages/page2.dart
